fix : refactor repository layer

+++ src/Applications/use_case/LogoutUserUseCase.js
@@ -1,5 +1,8 @@
 /* eslint-disable no-underscore-dangle */
 /* eslint-disable class-methods-use-this */
+const InvariantError = require('../../Commons/exceptions/InvariantError');
+const DomainErrorTranslator = require('../../Commons/exceptions/DomainErrorTranslator');
+
 class LogoutUserUseCase {
   constructor({
     authenticationRepository,
@@ -10,7 +13,14 @@ class LogoutUserUseCase {
   async execute(useCasePayload) {
     this._validatePayload(useCasePayload);
     const { refreshToken } = useCasePayload;
-    await this._authenticationRepository.checkAvailabilityToken(refreshToken);
+
+    const tokenCount = await this._authenticationRepository.checkAvailabilityToken(refreshToken);
+    if (tokenCount === 0) {
+      throw new InvariantError(
+        DomainErrorTranslator.translate(new Error('AUTHENTICATION.NOT_FOUND')).message,
+      );
+    }
+
     await this._authenticationRepository.deleteToken(refreshToken);
   }
